feat(datagrid): starting to move story files
